+++ query/plan/ir/VariableDependency.cpp
@@ -1,0 +1,11 @@
+#include "VariableDependency.h"
+
+using namespace db;
+
+void VariableDependency::addIncoming(DependencyEdge* newEdge) {
+    _incoming.push_back(newEdge);
+}
+
+void VariableDependency::addOutgoing(DependencyEdge* newEdge) {
+    _outgoing.push_back(newEdge);
+}
