+++ query/plan/ir/VariableDependencyGraph.h
@@ -1,126 +1,19 @@
 #pragma once
 
-#include <cstdint>
-#include <stdint.h>
-
 #include <deque>
 #include <string>
 #include <string_view>
-#include <type_traits>
 #include <unordered_map>
 #include <unordered_set>
 #include <vector>
 
-#include "EnumToString.h"
+#include "DependencyEdge.h"
+#include "VariableDependency.h"
 
 namespace db {
 
 class EntityPattern;
 class PatternElement;
-class VariableDependency;
-
-class EdgeMetadata {
-public:
-    enum class EdgeType : uint8_t;
-
-    explicit EdgeMetadata(EdgeType type)
-        : _type(type)
-    {
-    }
-
-    enum class EdgeType : uint8_t {
-        OUTGOING,
-        INCOMING,
-        BIDIRECTIONAL,
-        MERGE,
-
-        _SIZE
-    };
-
-    EdgeType type() const { return _type; }
-
-    static bool isMetaEdge(EdgeType et) { return et == EdgeType::MERGE; }
-
-    bool operator==(const EdgeMetadata& other) const {
-        return _type == other._type;
-    }
-
-private:
-    EdgeType _type {EdgeType::_SIZE};
-};
-
-static_assert(std::is_trivially_copyable_v<EdgeMetadata>);
-
-using EdgeTypeName = EnumToString<EdgeMetadata::EdgeType>::Create<
-    EnumStringPair<EdgeMetadata::EdgeType::OUTGOING, "getout">,
-    EnumStringPair<EdgeMetadata::EdgeType::INCOMING, "getin">,
-    EnumStringPair<EdgeMetadata::EdgeType::BIDIRECTIONAL, "bidir">,
-    EnumStringPair<EdgeMetadata::EdgeType::MERGE, "merge">
->;
-
-class DependencyEdge {
-public:
-    DependencyEdge(VariableDependency* src, VariableDependency* tgt, EdgeMetadata data)
-        : _src(src),
-         _tgt(tgt),
-        _data(data)
-    {
-    }
-
-    const VariableDependency* src() const { return _src; }
-    const VariableDependency* tgt() const { return _tgt; }
-    EdgeMetadata data() const { return _data; }
-
-    bool operator==(const DependencyEdge& other) const {
-        return _src == other.src() && _tgt == other.tgt() && _data == other.data();
-    }
-
-private:
-    friend class VariableDependencyGraph;
-
-    VariableDependency* _src;
-    VariableDependency* _tgt;
-    EdgeMetadata _data;
-};
-
-/**
- * @brief A representation of a variable and its dependencies.
- * @detail Used in @ref VariableDependencyGraph
- */
-class VariableDependency {
-public:
-    using Edges = std::vector<DependencyEdge*>;
-
-    explicit VariableDependency(std::string_view name)
-        : _name(name)
-    {
-    }
-
-    auto edges() const;
-
-    std::string_view getName() const { return _name; }
-
-    bool isIsolated() const;
-
-    void setName(std::string_view name) { _name = name; }
-
-    bool isSink() const { return _outgoing.empty(); }
-    bool isSource() const { return _incoming.empty(); }
-
-    const Edges& outgoing() const { return _outgoing; }
-    const Edges& incoming() const { return _incoming; }
-
-    void addIncoming(DependencyEdge* newEdge);
-    void addOutgoing(DependencyEdge* newEdge);
-
-private:
-    friend class VariableDependencyGraph;
-
-    std::string _name;
-
-    Edges _incoming;
-    Edges _outgoing;
-};
 
 /**
  * @brief Graph representation of the dependencies among variables, generated from
@@ -131,7 +24,7 @@ private:
 class VariableDependencyGraph {
 public:
     using Cycle = std::vector<VariableDependency*>;
-    
+
     /// Given a pattern (e.g. (n)-[e]->(m)), inserts all vars into the dependency graph
     void registerPatternElement(const PatternElement* ptn);
